+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Content Security Policy for configured Matomo servers in backend for TYPO3 v12 (#45)
+
 ## [2.1.1] - 2023-04-01
 
 ### Fixed